refactor: extract list projection helpers from resumes.ts to resumes_list_projections.ts
- Extract 19 functions and 8 type definitions (657 lines) to lib/resumes_list_projections.ts
- Includes: projectResumeListDoc, projectResumeDetailDoc, matchesResumeListFilters,
  sortResumeDocs, normalizeResumeListFilters, getIngestRuleScore, and all helper functions
- Add backward-compatible re-exports in resumes.ts
- Add 17 unit tests for the extracted module
- resumes.ts reduced from 3091 to 2482 lines (20% reduction)
